Renaming "base" package to solutions.cloudstark.quarkus.zalando.problem
The extension was renamed from "Quarkus problem" to "Quarkus Zalando problem". The package structure is now reflecting this change.
