+++ runtime/src/main/java/solutions/cloudstark/quarkus/zalando/problem/runtime/MediaType.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package solutions.cloudstark.quarkus.problem.runtime;
+package solutions.cloudstark.quarkus.zalando.problem.runtime;
 
 public class MediaType {
