+++ runtime/src/main/java/solutions/cloudstark/quarkus/zalando/problem/runtime/ThrowableProblemMapper.java
@@ -1,4 +1,4 @@
-package solutions.cloudstark.quarkus.problem.runtime;
+package solutions.cloudstark.quarkus.zalando.problem.runtime;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
