+++ runtime/src/main/java/solutions/cloudstark/quarkus/zalando/problem/runtime/jsonb/ConstraintViolationProblemSerializer.java
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package solutions.cloudstark.quarkus.problem.runtime.jsonb;
+package solutions.cloudstark.quarkus.zalando.problem.runtime.jsonb;
 
 import javax.json.bind.serializer.JsonbSerializer;
 import javax.json.bind.serializer.SerializationContext;
